+++ Sheepskin/Library/Data/Instrument/ETF/ETF.h
@@ -11,7 +11,7 @@
 class ETF: public Instrument {
 public:
     explicit ETF(std::string id);
-    long double getVolatility(const Record &record) override;
+    double getVolatility(const Record &record) override;
     double getPrice(const Record &record) override;
 };
