+++ Sheepskin/Library/Data/Instrument/Stock/Stock.h
@@ -12,7 +12,7 @@
 class Stock: public Instrument {
 public:
         explicit Stock(std::string id);
-        long double getVolatility(const Record &record) override;
+        double getVolatility(const Record &record) override;
         double getPrice(const Record &record) override;
         double calculateMean(const Record &record);
 };
